Implement scene space line strip picking
Implement picking with a line strip in scene space. This will allow VR
users to pick into the scene from their VR controller like a laser pen
rather than only from the camera view, and also allow for picking with a
parabolic arc as is commonly used to implement teleportation.