Implement scene space line strip picking
Implement picking with a line strip in scene space, rather than only
from the camera view with mouse coordinates. This will allow for VR
users to pick into the scene from their head position for 360 mouse
mode, to pick from their VR controller like a laser pen, and also
picking with a parabolic arc as is commonly used for teleportation.