Implement a 3D mouse cursor
Implement a new FGMouseCursor3D class, which represents a 3D mouse
cursor in the scene graph. This will be used to make the normal mouse
pointer visible to VR users, and also to allow for VR controller based
mouse emulation in future.

The class places itself in the scene graph under the aircraft transform
so that it doesn't flicker around on aircraft motion due to update
ordering.