i18n: add four classes to prepare (mainly) for plural forms handling
Add classes TranslationDomain, TranslationResource, TranslationUnit and
DefaultTranslationParser.

We want to have plural forms grouped together, however the way
readProperties() reads linguistically unrelated sibling nodes such as
the <fg-scenery-desc> nodes from Translations/default/options.xml gets
in the way. We might be able to postprocess the resulting property tree
to “separate” such siblings, but that would be an ugly hack. Instead, we
choose to load related source texts and target texts into dedicated
TranslationUnit instances; we load them from XML files in
$FG_ROOT/Translations/default/ which are very close to the previously
used files, but are not PropertyList files anymore.

DefaultTranslationParser is the class for reading these new XML files.
It handles a 'has-plural' attribute to declare source texts whose
corresponding translations may have different plural forms (users of the
API will pass an integer to
GetLocalizedStringsSetup::setCardinalNumber(), which will allow for
instance GetLocalizedStringsSetup::getString() to select the appropriate
plural form for this number of items, or whatever it is).