FGLocale: add _languageId member and getLanguageId() getter
We need the _languageId value every time a plural form is needed, and
every time we want to find the number of plural forms in the selected
locale.

With this commit, FGLocale::selectLanguage() performs all the necessary
checks, logs alert messages as appropriate and sets _languageId so that
the rest of the code can focus on its own business and just use
getLanguageId() without having to duplicate all the checks and logging.