Marker pins: Convert QUAD_STRIP to TRANGLE_FAN
Marker pin geometry is using QUAD_STRIP, which is deprecated in the
OpenGL core profile. It is also a bit redundant as the lower row of
vertices are all the same point, so a TRIANGLE_FAN is a much more
natural fit.

Changes are still needed to Effects/marker-pin to allow them to render
with the HDR compositor.