FGLocale: load files from the new 'auto-extracted' subdirectory
The new 'fg-extract-translatable-strings' script in FGMeta extracts
translatable strings from XML files in FGData, aircraft or add-ons that
have the special 'FlightGear-tr' processing instruction. When used in
'write' mode, this script writes the result in default translation
format (XML) to a subdirectory named 'auto-extracted' of
Translations/default (this way, the script can remove stale files from
previous extractions without resorting to magic). This commit causes
FlightGear to load such files, so that the extracted strings can be
used.

These files are only a part of the default translation. In order to have
actual translations, one of course has to propagate the extracted
strings to the usual FlightGear-nonQt.xlf files and translate these
files, as usual (fg-update-translation-files handles the "propagation"
step and will read strings from the new 'auto-extracted' subdirectory
once [1] is merged).

With merge request [1] and this commit, the FGTranslate class can
provide translations of automatically-extracted strings from FGData,
aircraft or add-ons.

[1] https://gitlab.com/flightgear/fgmeta/-/merge_requests/55