FGTranslate: API and implementation changes
The main goal here is to allow one to efficiently perform several
operations on the same TranslationUnit, such as:
  - querying whether it has plural forms before getting a translation
    (this should be used by future code for <format> handling in GUI
    XML dialogs);
  - retrieving translations for different values of the “cardinal
    number”.

In TranslationResource, the mapped_type of the _map data member is now
std::shared_ptr<TranslationUnit>.

Add functions FGTranslate::translationUnit() that returns copies of such
shared pointers, TranslationUnit::getTranslation() and
TranslationUnit::getTranslation(intType cardinalNumber) that return
translations (strings): the former for strings without plural forms, the
latter for those with plural forms.

Change the _domains map in FGLocale to contain shared pointers and
modify FGTranslate to store as data member a shared pointer to the
TranslationDomain rather than the domain as a string. This will save a
bunch of lookups in the _domains map. Add an FGTranslate constructor
that immediately sets the _domain data member.

In FGTranslate, don't store the index and cardinal number as data
members anymore (the only remaining data member is a shared pointer to
the translation domain). The index is now an optional argument of get()
and getWithDefault(). Add two analogous functions getPlural() and
getPluralWithDefault() that require an additional argument: they are for
use with strings that have plural forms.

TranslationResource: remove two unused member functions.
TranslationDomain: rename getResourceCreate() to getOrCreateResource().