PUICompatObject: allow translations whose plural status is determined at runtime
This is for the <format> string of some widgets such as XMLLabel. When
such strings are defined with translate="y" in the dialog XML config, we
want special handling that only calls a plural-enabled translation
function if the translatable string was defined with has-plural="true"
*and* the "local value" is an integer (which is then the "cardinal
number" that determines the appropriate plural form).

This goes with the following FGData commit:

  XMLDialog.nas: enable localized <format> strings with plural forms