View: Expose VR transformed view
The View class is not currently aware of the additional VR head pose
transform applied to the view position & orientation. Apply the
additional transform (based of the view reference space exposed in the
latest osgXR 0.5.6) so that FGSoundManager picks up the accurate
position and orientation of the VR head pose, but also add additional
accessors which exclude the VR head pose for passing to the camera
group, since that already hands the individual VR view transforms.

This makes FGSoundManager position and orient sounds relative to the
head, which is great for headphones. It will also allow a 3D mouse
cursor to move relative to the current head pose.