FGRenderer: Treat first physical object as hit
Make FGRenderer::pick() treat the first physical object encountered as a
hit, even without a pick callback. This will allow the upcoming 3D mouse
cursor to be visible on the surface of objects that don't have
callbacks.

FGMouseInput needs some extra checks so as not to crash when it
encounters scenery picks without callbacks.