Implement 360 mouse
Implement 360 mouse capability in FGMouseInput and FGMouseCursor3D,
intended for use with VR. The 360 mouse cursor is enabled when the mouse
mode's 'vr-cursor/pass-through' node is set, in which case the normal mouse
cursor is hidden and mouse events pass through to wherever the 3D mouse
cursor is targetted.

FGMouseInput listens to /sim/vr/cursors/mouse[0]/dx & -dy, which can be
set from mouse mode axis bindings to move the 360 mouse cursor target.
Motion is relative to the view pose and can wrap right around. The
target is only updated on explicit user action such as mouse events (so
lateral head motion only temporarily pops the 3D mouse cursor forward if
the target is occluded).