CameraGroup: Fix edge case in viewport intersection
CameraGroup's computeCameraIntersection() receives floating point window
coordinates calculated by eventToWindowCoords(). These are in the range
0.0..(double)traits->width (or ->height). As such, at the extreme edge
of the window they contain coordinate values matching the width and
height of the graphics context.

However computeCameraIntersection() tests whether they are contained
within the viewport described by an SGRect<double> with intended maximum
(x + w - 1.0, y + h - 1.0). In fact the SGRect constructor actually
takes a width and height, not max_x, max_y, so for non-zero x & y it
will be even further off. Usually (with zero x & y) this results in the
top row and right column of pixels in the window not matching the
viewport and mouse move events being delivered to the wrong camera. Fix
this SGRect to span all the way to (x + w, y + h).

This is particularly visible with the upcoming 3D mouse cursor enabled,
as it freezes the 3D cursor despite horizontal motion *before*
disappearing when leaving the graphics area, and with the VR in use it
jumps position before disappearing due to matching the larger VR camera
which is also in the CameraGroup.