i18n: fixes in DefaultTranslationParser and TranslationResource
In DefaultTranslationParser::endElement(): remove one std::move() that
is quite likely to ruin _stringTagName in the expression
_nextIndex[_stringTagName]++ (this may lead to large indices if the
std::move() causes the same string value in std::map<>::operator[] to be
used for many translatable strings).

In TranslationResource::setFirstTargetText(): don't add
default-constructed elements to the map when the key isn't found. The
map is supposed to only contain elements for (basicId, index) pairs that
exist in the default translation (TranslationResource::getAll() and
TranslationResource::getCount() rely on this fact).