MouseCursor: Fix StockOSGCursor::setCursorVisible
The StockOSGCursor class implements setCursorVisible() with a quick
return if mCursorObscured == aVis, however this appears to be a
copy/paste error as setCursorVisible() only changes mCursorVisible,
which has the opposite polarity to mCursorObscured.

Correct the condition to test mCursorVisible to allow the mouse cursor
to be hidden even if not already obscured via
hideCursorUntilMouseMove().