Implement a 3D direction cue
Add an FGDirectionCue3D class to represent a direction cue in the 3D
scene, pointing towards another node in the scene. This will be used for
pointing at the 3D mouse cursor when it isn't in view, especially useful
for VR where you can more naturally turn your head away from the cursor.
It could also potentially be adapted for use with the tutorials instead
of forcefully turning the view (which could be nausiating in VR).

The 3D model is loaded from /sim/vr/config/cursors/cue/model/path.