GraphicsPresets: fix support for translated names and descriptions
A preset name and description can be translated if the preset
PropertyList declares a <translation-domain> (one may also specify a
<translation-context>, however this is optional as the translation
context can be derived from the preset file name). If this is the case
and the <name> and <description> nodes have translate="y", then
translations of the preset name and description are used.

This implements #3170 and removes the following warning that appeared
when opening the Rendering Options dialog:
TranslationDomain::getResource(): unable to find requested resource
'graphics-presets'.