splash: Fix dynamic text & leak
Splash text should be marked with a DataVariance of DYNAMIC if
dynamicContent is set, otherwise the vertex array object won't be set
again and the text won't update. However any use of DYNAMIC DataVariance
(which is currently in use in multiple places in FlightGear) completely
prevents multithreaded frame pipelining, so lets keep it static and just
create a new copy of the osgText::Text object if the text changes.

To be able to replace the old Text object, we need to access its parent.
However when the splash screen is hidden, the splash screen's child
nodes are removed, orphaning the text nodes which are still referenced
in _items. These stale parentless text nodes accumulate each time the
splash is made visible again.

Plug the leak when hiding the splash by clearing _items alongside the
removal of all child nodes, allowing the dynamic text update code to
assume the text node has a parent (which we make explicit with an
assert).