Remove terrasync.py from this repository
terrasync.py (with its modules and automated tests) is now in the
fgmeta-python repository[1]. It is installable along with other tools
for FlightGear development via a Python package called 'flightgear'
(see [2] or scroll down at [1] for instructions and details). This
move allowed us to share and enhance the DirIndex class for use in tools
that prepare the base package (create_dirindex.py, create_base_package.sh).

[1] https://gitlab.com/flightgear/fgmeta-python
[2] https://gitlab.com/flightgear/fgmeta-python/-/blob/next/README.md?ref_type=heads